(MODULES-7661) Fix JSON encoding for single object arrays
Previously the powershell tests were failing on tests where they returned a
single JSON object.  This is due to percularity in PowerShell where single
element arrays get reduced to just the item.  This commit changes the JSON
encoding to insert a begin and end bracket into the JSON string which forces the
output to be an array again.  This only applies for single elements not for
empty or multi element arrays.