(GH-CAT-9) Test file update
Prior to this commit, a different commit containing some syntax update
was implemented. However, this update broke several test cases.

This commit aims to update the unit_spec.rb test file so that its test
cases keep working as intended. It also makes a few changes on top of
the previous syntax update to help the test file run correctly.